test: split vim tests to mirror the module layout
